fix(engine): win64(LLP64) 加固 ScriptVariant 对齐与 filestream 尺寸类型
静态扫描 A/B/C 三类后落地的两处 win64 正确性加固:
- ScriptVariant.h: 去掉 #pragma pack(4)。该 union 含 8 字节指针 ptrVal,
  win64 下 pack(4) 破坏其自然对齐/取整; 结构纯内存态、不参与二进制序列化,
  去 pack 无布局回归。
- openborscript.c load_file_stream: long size -> size_t, 与 s_filestream.size
  及 buffer_pakfile(size_t*) 一致, 消除 win64 下 ftell 结果经 long(4B) 中转进
  malloc/fread 的截断/符号扩展。

A 类(font_printf %lu + 指针混用)经复核实参均为 unsigned, win64 下与 %lu 同宽,
恰匹配, 改 %llu 反致越界读, 故不改。实测本组改动后 win64 仍 0xC0000374,
这两处为加固非致命点; 精确崩溃点仍需运行时栈。

+++ engine/openborscript.c
@@ -9499,7 +9499,9 @@ HRESULT openbor_openfilestream(ScriptVariant **varlist , ScriptVariant **pretvar
     FILE *handle = NULL;
     char path[MAX_BUFFER_LEN] = {""};
     char tmpname[MAX_BUFFER_LEN] = {""};
-    long size;
+    size_t size;   // win64(LLP64): 原 long 仅 4 字节, ftell 结果经其中转再进
+                   // malloc/fread(size_t) 会截断/符号扩展; 与 s_filestream.size
+                   // 及 buffer_pakfile(size_t*) 统一为 size_t。
 
     ScriptVariant_ChangeType(*pretvar, VT_INTEGER);
